Extract server side JsonRpcFunction & handler creation
