+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/server/JsonRpcHandlerMacroFactory.scala
@@ -1,0 +1,131 @@
+package io.github.shogowada.scala.jsonrpc.server
+
+import io.github.shogowada.scala.jsonrpc.Models.{JsonRpcError, JsonRpcRequest}
+import io.github.shogowada.scala.jsonrpc.server.JsonRpcServer.Handler
+import io.github.shogowada.scala.jsonrpc.utils.MacroUtils
+
+import scala.reflect.macros.blackbox
+
+class JsonRpcHandlerMacroFactory[CONTEXT <: blackbox.Context](val c: CONTEXT) {
+
+  import c.universe._
+
+  lazy val functionMacroFactory = new JsonRpcFunctionMacroFactory[c.type](c)
+
+  def createHandler[API](
+      server: c.Tree,
+      maybeClient: Option[c.Tree],
+      api: c.Expr[API],
+      method: c.universe.MethodSymbol
+  ): c.Expr[Handler] = {
+    val parameterTypeLists: List[List[Type]] = method.asMethod.paramLists
+        .map(parameters => parameters.map(parameter => parameter.typeSignature))
+
+    createHandler(
+      server,
+      maybeClient,
+      q"$api.$method",
+      parameterTypeLists,
+      method.asMethod.returnType
+    )
+  }
+
+  def createHandler(
+      server: c.Tree,
+      maybeClient: Option[c.Tree],
+      method: c.Tree,
+      parameterTypeLists: Seq[Seq[c.Type]],
+      returnType: c.Type
+  ): c.Expr[Handler] = {
+    val macroUtils = MacroUtils[c.type](c)
+
+    val jsonSerializer = macroUtils.getJsonSerializer(server)
+    val executionContext = macroUtils.getExecutionContext(server)
+
+    val parameterTypes: Seq[Type] = parameterTypeLists.flatten
+
+    val jsonRpcParameterType: Tree = macroUtils.getJsonRpcParameterType(parameterTypes)
+
+    def arguments(params: TermName): Seq[Tree] = {
+      parameterTypes.indices
+          .map(index => {
+            val parameterType = parameterTypes(index)
+            val fieldName = TermName(s"_${index + 1}")
+            val argument = q"$params.$fieldName"
+            if (macroUtils.isJsonRpcFunctionType(parameterType)) {
+              maybeClient
+                  .map(client => functionMacroFactory.getOrCreateJsonRpcFunction(server, client, parameterType, argument))
+                  .getOrElse(throw new UnsupportedOperationException("To use JsonRpcFunction, you need to bind the API to JsonRpcServerAndClient."))
+            } else {
+              argument
+            }
+          })
+    }
+
+    val json = TermName("json")
+    val params = TermName("params")
+
+    def methodInvocation(params: TermName) = {
+      if (parameterTypeLists.isEmpty) {
+        q"$method"
+      } else {
+        q"$method(..${arguments(params)})"
+      }
+    }
+
+    def notificationHandler = c.Expr[Handler](
+      q"""
+          ($json: String) => {
+            ..${macroUtils.imports}
+            $jsonSerializer.deserialize[JsonRpcNotification[$jsonRpcParameterType]]($json)
+              .foreach(notification => {
+                val $params = notification.params
+                ${methodInvocation(params)}
+              })
+            Future(None)($executionContext)
+          }
+          """
+    )
+
+    def requestHandler = {
+      val request = TermName("request")
+
+      val maybeInvalidParamsErrorJson: c.Expr[Option[String]] =
+        macroUtils.createMaybeErrorJson(
+          server,
+          c.Expr[String](q"$json"),
+          c.Expr[JsonRpcError[String]](q"JsonRpcErrors.invalidParams")
+        )
+
+      def maybeJsonRpcRequest(json: TermName) = c.Expr[Option[JsonRpcRequest[jsonRpcParameterType.type]]](
+        q"""$jsonSerializer.deserialize[JsonRpcRequest[$jsonRpcParameterType]]($json)"""
+      )
+
+      c.Expr[Handler](
+        q"""
+            ($json: String) => {
+              ..${macroUtils.imports}
+              ${maybeJsonRpcRequest(json)}
+                .map(($request: JsonRpcRequest[$jsonRpcParameterType]) => {
+                  val $params = $request.params
+                  ${methodInvocation(params)}
+                    .map((result) => JsonRpcResultResponse(
+                      jsonrpc = Constants.JsonRpc,
+                      id = $request.id,
+                      result = result
+                    ))($executionContext)
+                    .map((response) => $jsonSerializer.serialize(response))($executionContext)
+                })
+                .getOrElse(Future($maybeInvalidParamsErrorJson)($executionContext))
+            }
+            """
+      )
+    }
+
+    if (macroUtils.isJsonRpcNotificationMethod(returnType)) {
+      notificationHandler
+    } else {
+      requestHandler
+    }
+  }
+}
