+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/server/JsonRpcServer.scala
@@ -1,6 +1,6 @@
 package io.github.shogowada.scala.jsonrpc.server
 
-import io.github.shogowada.scala.jsonrpc.Models.{JsonRpcError, JsonRpcRequest}
+import io.github.shogowada.scala.jsonrpc.Models.JsonRpcError
 import io.github.shogowada.scala.jsonrpc.serializers.JsonSerializer
 import io.github.shogowada.scala.jsonrpc.server.JsonRpcServer.Handler
 import io.github.shogowada.scala.jsonrpc.utils.MacroUtils
@@ -77,174 +77,12 @@ object JsonRpcServerMacro {
     import c.universe._
 
     val macroUtils = MacroUtils[c.type](c)
+    val handlerMacroFactory = new JsonRpcHandlerMacroFactory[c.type](c)
 
     val methodName = macroUtils.getJsonRpcMethodName(method)
-    val handler = createHandler[c.type, API](c)(server, maybeClient, api, method)
+    val handler = handlerMacroFactory.createHandler[API](server, maybeClient, api, method)
 
     c.Expr[(String, Handler)](q"""$methodName -> $handler""")
-  }
-
-  def createHandler[CONTEXT <: blackbox.Context, API](c: CONTEXT)(
-      server: c.Tree,
-      maybeClient: Option[c.Tree],
-      api: c.Expr[API],
-      method: c.universe.MethodSymbol
-  ): c.Expr[Handler] = {
-    import c.universe._
-
-    val parameterTypeLists: List[List[Type]] = method.asMethod.paramLists
-        .map(parameters => parameters.map(parameter => parameter.typeSignature))
-
-    createHandler[c.type](c)(
-      server,
-      maybeClient,
-      q"$api.$method",
-      parameterTypeLists,
-      method.asMethod.returnType
-    )
-  }
-
-  def createHandler[CONTEXT <: blackbox.Context](c: CONTEXT)(
-      server: c.Tree,
-      maybeClient: Option[c.Tree],
-      method: c.Tree,
-      parameterTypeLists: Seq[Seq[c.Type]],
-      returnType: c.Type
-  ): c.Expr[Handler] = {
-    import c.universe._
-
-    val macroUtils = MacroUtils[c.type](c)
-
-    val jsonSerializer = macroUtils.getJsonSerializer(server)
-    val executionContext = macroUtils.getExecutionContext(server)
-
-    val parameterTypes: Seq[Type] = parameterTypeLists.flatten
-
-    val jsonRpcParameterType: Tree = macroUtils.getJsonRpcParameterType(parameterTypes)
-
-    def arguments(params: TermName): Seq[Tree] = {
-      parameterTypes.indices
-          .map(index => {
-            val parameterType = parameterTypes(index)
-            val fieldName = TermName(s"_${index + 1}")
-            val argument = q"$params.$fieldName"
-            if (macroUtils.isJsonRpcFunctionType(parameterType)) {
-              maybeClient
-                  .map(client => getOrCreateJsonRpcFunction[c.type](c)(server, client, parameterType, argument))
-                  .getOrElse(throw new UnsupportedOperationException("To use JsonRpcFunction, you need to bind the API to JsonRpcServerAndClient."))
-            } else {
-              argument
-            }
-          })
-    }
-
-    val json = TermName("json")
-    val params = TermName("params")
-
-    def methodInvocation(params: TermName) = {
-      if (parameterTypeLists.isEmpty) {
-        q"$method"
-      } else {
-        q"$method(..${arguments(params)})"
-      }
-    }
-
-    def notificationHandler = c.Expr[Handler](
-      q"""
-          ($json: String) => {
-            ..${macroUtils.imports}
-            $jsonSerializer.deserialize[JsonRpcNotification[$jsonRpcParameterType]]($json)
-              .foreach(notification => {
-                val $params = notification.params
-                ${methodInvocation(params)}
-              })
-            Future(None)($executionContext)
-          }
-          """
-    )
-
-    def requestHandler = {
-      val request = TermName("request")
-
-      val maybeInvalidParamsErrorJson: c.Expr[Option[String]] =
-        macroUtils.createMaybeErrorJson(
-          server,
-          c.Expr[String](q"$json"),
-          c.Expr[JsonRpcError[String]](q"JsonRpcErrors.invalidParams")
-        )
-
-      def maybeJsonRpcRequest(json: TermName) = c.Expr[Option[JsonRpcRequest[jsonRpcParameterType.type]]](
-        q"""$jsonSerializer.deserialize[JsonRpcRequest[$jsonRpcParameterType]]($json)"""
-      )
-
-      c.Expr[Handler](
-        q"""
-            ($json: String) => {
-              ..${macroUtils.imports}
-              ${maybeJsonRpcRequest(json)}
-                .map(($request: JsonRpcRequest[$jsonRpcParameterType]) => {
-                  val $params = $request.params
-                  ${methodInvocation(params)}
-                    .map((result) => JsonRpcResultResponse(
-                      jsonrpc = Constants.JsonRpc,
-                      id = $request.id,
-                      result = result
-                    ))($executionContext)
-                    .map((response) => $jsonSerializer.serialize(response))($executionContext)
-                })
-                .getOrElse(Future($maybeInvalidParamsErrorJson)($executionContext))
-            }
-            """
-      )
-    }
-
-    if (macroUtils.isJsonRpcNotificationMethod(returnType)) {
-      notificationHandler
-    } else {
-      requestHandler
-    }
-  }
-
-  def getOrCreateJsonRpcFunction[CONTEXT <: blackbox.Context](c: CONTEXT)(
-      server: c.Tree,
-      client: c.Tree,
-      jsonRpcFunctionType: c.Type,
-      jsonRpcFunctionMethodName: c.Tree
-  ): c.Tree = {
-    import c.universe._
-
-    val newJsonRpcFunction = createJsonRpcFunction[c.type](c)(server, client, jsonRpcFunctionType, jsonRpcFunctionMethodName)
-
-    q"""
-        $newJsonRpcFunction
-        """
-  }
-
-  def createJsonRpcFunction[CONTEXT <: blackbox.Context](c: CONTEXT)(
-      server: c.Tree,
-      client: c.Tree,
-      jsonRpcFunctionType: c.Type,
-      jsonRpcFunctionMethodName: c.Tree
-  ): c.Tree = {
-    import c.universe._
-
-    val macroUtils = MacroUtils[c.type](c)
-
-    val functionType: Type = macroUtils.getFunctionTypeOfJsonRpcFunctionType(jsonRpcFunctionType)
-    val functionTypeTypeArgs: Seq[Type] = functionType.typeArgs
-    val paramTypes: Seq[Type] = functionTypeTypeArgs.init
-    val returnType: Type = functionTypeTypeArgs.last
-    val function = macroUtils.createClientMethodAsFunction(client, Some(server), jsonRpcFunctionMethodName, paramTypes, returnType)
-
-    q"""
-        new {} with JsonRpcFunction[$functionType] {
-          override val function = $function
-
-          override def dispose(): Future[Unit] = {
-            Future.failed(new UnsupportedOperationException("TODO: dispose the function"))
-          }
-        }
-        """
   }
 
   def receive(c: blackbox.Context)(json: c.Expr[String]): c.Expr[Future[Option[String]]] = {
